Refactor logic into separate modules
