+++ README.md
@@ -1,1 +1,12 @@
 # bodybest
+
+Това хранилище съдържа клиентския код на демо таблото **MyBody.Best**.
+
+## Структура на JavaScript модулите
+
+- `app.js` – основен файл, който инициализира приложението и съхранява глобалното състояние.
+- `dataLoader.js` – отговаря за зареждането на данните на таблото.
+- `chat.js` – управление на чата и изпращане на съобщения към сървъра.
+- `quizProcessing.js` – логика за генериране и подаване на адаптивния въпросник.
+
+Всички скриптове се намират в папка `js/` и се използват като ES модули.
