+++ js/app.js
@@ -15,19 +15,9 @@ import { populateUI } from './populateUI.js';
 // КОРЕКЦИЯ: Премахваме handleDelegatedClicks от импорта тук
 import { setupStaticEventListeners, setupDynamicEventListeners } from './eventListeners.js';
 import { handleLogout as performLogout } from './auth.js';
-import {
-    toggleChatWidget, closeChatWidget, displayMessage as displayChatMessage,
-    displayTypingIndicator as displayChatTypingIndicator, scrollToChatBottom
-} from './chat.js';
+import { toggleChatWidget, closeChatWidget } from './chat.js';
+import { loadDashboardData } from './dataLoader.js';
 import { openExtraMealModal } from './extraMealForm.js';
-import {
-    openAdaptiveQuizModal as _openAdaptiveQuizModal,
-    renderCurrentQuizQuestion,
-    showQuizValidationMessage,
-    hideQuizValidationMessage,
-    getSummaryFromLastCompletedQuiz, // For potential use in app.js
-    getSummaryFromPreviousQuizzes // For potential use in app.js
-} from './adaptiveQuiz.js';
 
 // ==========================================================================
 // ГЛОБАЛНИ ПРОМЕНЛИВИ ЗА СЪСТОЯНИЕТО НА ПРИЛОЖЕНИЕТО
@@ -63,25 +53,6 @@ export function resetAppState() {
     // Не нулираме toastTimeoutApp, тъй като showToast управлява своя.
 }
 
-// Функция за създаване на тестови данни
-function createTestData() {
-    return {
-        success: true,
-        planStatus: "ready",
-        userName: "Тестов Потребител",
-        showAdaptiveQuiz: false,
-        analytics: {
-            current: { goalProgress: 65, engagementScore: 78, overallHealthScore: 72 },
-            detailed: [{ label: "BMI", initialValueText: "28.5", expectedValueText: "24.0", currentValueText: "26.2", key: "bmi", infoTextKey: "bmi_info" }],
-            textualAnalysis: "Вашият напредък е стабилен. Продължавайте със силната работа!"
-        },
-        planData: { week1Menu: { monday: [{ meal_name: "Закуска", items: [{ name: "Овесена каша", grams: "50г" }, { name: "Банан", grams: "1 бр." }] }] }},
-        dailyLogs: [],
-        currentStatus: { weight: 75.5 },
-        initialData: { weight: 80.0, height: 175 },
-        initialAnswers: { goal: "отслабване", age: "30", gender: "жена" }
-    };
-}
 
 
 // ==========================================================================
@@ -124,128 +95,6 @@ function initializeApp() {
 // ==========================================================================
 // ЗАРЕЖДАНЕ И ОБРАБОТКА НА ДАННИ
 // ==========================================================================
-export async function loadDashboardData() { // Exported for adaptiveQuiz.js to call after submit
-    console.log("loadDashboardData starting for user:", currentUserId);
-    if (!currentUserId) {
-         showPlanPendingState("Грешка: Потребителска сесия не е намерена. Моля, <a href='index.html' style='color: var(--primary-color); text-decoration: underline;'>влезте отново</a>.");
-         showLoading(false);
-         return;
-    }
-    showLoading(true, "Зареждане на вашето персонализирано табло...");
-    Object.keys(todaysMealCompletionStatus).forEach(key => delete todaysMealCompletionStatus[key]); // Clear previous
-
-    try {
-        if (currentUserId.includes('test_user') || window.location.hostname.includes('replit')) {
-            const data = createTestData();
-            console.log("Using test data for development:", data);
-            fullDashboardData = data;
-            chatHistory = []; // Reset chat history for test user on reload
-
-            if(selectors.planPendingState) selectors.planPendingState.classList.add('hidden');
-            if(selectors.appWrapper) selectors.appWrapper.style.display = 'block';
-
-            populateUI();
-            setupDynamicEventListeners();
-
-            const activeTabId = sessionStorage.getItem('activeTabId') || selectors.tabButtons[0]?.id;
-            const activeTabButton = activeTabId ? document.getElementById(activeTabId) : (selectors.tabButtons && selectors.tabButtons[0]);
-
-            if (activeTabButton && Array.from(selectors.tabButtons).includes(activeTabButton)) {
-                activateTab(activeTabButton);
-            } else if (selectors.tabButtons && selectors.tabButtons.length > 0) {
-                activateTab(selectors.tabButtons[0]);
-            }
-
-            showToast("Тестов режим активиран успешно.", false, 2000);
-            showLoading(false);
-            return;
-        }
-
-        const response = await fetch(`${apiEndpoints.dashboard}?userId=${currentUserId}`);
-        if (!response.ok) throw new Error(`Грешка от сървъра: ${response.status} ${response.statusText}`);
-        const data = await response.json();
-        if (!data.success) throw new Error(data.message || 'Неуспешно зареждане на данни от сървъра.');
-
-        console.log("Data received from worker:", data);
-        fullDashboardData = data;
-        // chatHistory = []; // Do not reset chat history on normal data load, only for test user or logout
-
-        if (data.planStatus === "pending" || data.planStatus === "processing") {
-            showPlanPendingState(); return;
-        }
-        if (data.planStatus === "error") {
-            showPlanPendingState(`Възникна грешка при генерирането на вашия план: ${data.message || 'Свържете се с поддръжка.'}`); return;
-        }
-
-        if(selectors.planPendingState) selectors.planPendingState.classList.add('hidden');
-        if(selectors.appWrapper) selectors.appWrapper.style.display = 'block';
-
-        const welcomeSeenKey = `welcomeScreenSeen_${currentUserId}`;
-        if (data.planStatus === "ready" && data.isFirstLoginWithReadyPlan && !localStorage.getItem(welcomeSeenKey)) {
-            localStorage.setItem(welcomeSeenKey, 'true');
-            openModal('welcomeScreenModal');
-        }
-
-        if (data.aiUpdateSummary) {
-            const { title, introduction, changes, encouragement } = data.aiUpdateSummary;
-            let summaryHtml = `<h3>${title || 'Важни Актуализации'}</h3>`;
-            if (introduction) summaryHtml += `<p>${introduction.replace(/\n/g, '<br>')}</p>`;
-            if (changes && Array.isArray(changes) && changes.length > 0) {
-                summaryHtml += `<ul>${changes.map(ch => `<li>${String(ch).replace(/\n/g, '<br>')}</li>`).join('')}</ul>`;
-            }
-            if (encouragement) summaryHtml += `<p>${encouragement.replace(/\n/g, '<br>')}</p>`;
-
-            if (selectors.infoModalTitle) selectors.infoModalTitle.textContent = title || 'Информация';
-            if (selectors.infoModalBody) selectors.infoModalBody.innerHTML = summaryHtml;
-            openModal('infoModal');
-            fetch(apiEndpoints.acknowledgeAiUpdate, {
-                 method: 'POST',
-                 headers: {'Content-Type': 'application/json'},
-                 body: JSON.stringify({userId: currentUserId})
-            }).catch(err => console.warn("Failed to acknowledge AI update:", err));
-        }
-
-        populateUI();
-        setupDynamicEventListeners();
-
-        const activeTabId = sessionStorage.getItem('activeTabId') || selectors.tabButtons[0]?.id;
-        const activeTabButton = activeTabId ? document.getElementById(activeTabId) : (selectors.tabButtons && selectors.tabButtons[0]);
-
-        if (activeTabButton && Array.from(selectors.tabButtons).includes(activeTabButton)) {
-            activateTab(activeTabButton);
-        } else if (selectors.tabButtons && selectors.tabButtons.length > 0) {
-            activateTab(selectors.tabButtons[0]);
-        }
-
-        if (fullDashboardData.showAdaptiveQuiz === true) {
-            _openAdaptiveQuizModal(); // From adaptiveQuiz.js
-        }
-
-        showToast("Данните са заредени успешно.", false, 2000);
-    } catch (error) {
-        console.error("Error loading/processing dashboard data:", error);
-        showToast(`Грешка при зареждане: ${error.message}`, true, 7000);
-        showPlanPendingState(`Възникна грешка: ${error.message}. Опитайте да презаредите страницата или <a href='index.html' style='color: var(--primary-color); text-decoration: underline;'>влезте отново</a>.`);
-    } finally {
-        showLoading(false);
-    }
-}
-
-function showPlanPendingState(customMessage) {
-    if (selectors.appWrapper) selectors.appWrapper.style.display = 'none';
-    if (selectors.planPendingState) {
-        selectors.planPendingState.classList.remove('hidden');
-        const pElements = selectors.planPendingState.querySelectorAll('p');
-        if (customMessage) {
-            if (pElements.length > 1) pElements[1].innerHTML = customMessage;
-            else if (pElements.length > 0) pElements[0].innerHTML = customMessage;
-        } else {
-             if (pElements.length > 0) pElements[0].textContent = "Благодарим ви за попълнения въпросник! Вашият персонализиран план MyBody.Best се изготвя.";
-             if (pElements.length > 1) pElements[1].textContent = "Моля, проверете отново по-късно. Ще бъдете уведомени (ако сте позволили известия) или опитайте да презаредите страницата след известно време.";
-        }
-    }
-    showLoading(false);
-}
 
 // ==========================================================================
 // ФУНКЦИИ ЗА УПРАВЛЕНИЕ НА ДАННИ (ЗАПИС)
@@ -348,228 +197,6 @@ export async function handleFeedbackFormSubmit(event) { // Exported for eventLis
         selectors.feedbackForm.reset();
     }, 1500);
 }
-
-
-// ==========================================================================
-// АДАПТИВЕН ВЪПРОСНИК - ЛОГИКА И API КОМУНИКАЦИЯ (Остават в app.js)
-// ==========================================================================
-
-export async function _generateAdaptiveQuizClientSide(userId, context = {}) { // Prefixed to avoid clash if adaptiveQuiz.js also defines it
-    console.log("generateAdaptiveQuizClientSide (app.js) called for user:", userId, "with context:", context);
-    if (!userId) throw new Error("Липсва потребителско ID за генериране на въпросник.");
-
-    if (userId.includes('test_user') || window.location.hostname.includes('replit')) {
-        console.log("Generating test quiz data (app.js)");
-        return {
-            quizId: 'test_quiz_' + Date.now(),
-            quizTitle: "Тестов Въпросник от App.js",
-            quizDescription: "Това е тестов въпросник за разработка от App.js",
-            questions: [
-                { id: "q1_app", text: "Как се чувствате днес (App)?", answerType: "скала_1_5", required: true, options: { min: 1, max: 5, minLabel: "Зле", maxLabel: "Отлично" }},
-                { id: "q2_app", text: "Колко чаши вода пихте вчера (App)?", answerType: "number", required: true, placeholder: "Въведете брой чаши" },
-                { id: "q3_app", text: "Какви са главните ви предизвикателства (App)?", answerType: "еднозначен_избор_от_списък", required: false, options: ["Време", "Мотивация", "Знания", "Активност"] }
-            ]
-        };
-    }
-
-    try {
-        let queryString = `userId=${userId}`;
-        if (context.trigger) queryString += `&trigger=${encodeURIComponent(context.trigger)}`;
-        if (context.specificFocus) queryString += `&focus=${encodeURIComponent(context.specificFocus)}`;
-        const response = await fetch(`${apiEndpoints.getAdaptiveQuiz}?${queryString}`);
-        if (!response.ok) {
-            const errorData = await response.json().catch(() => ({ message: `Сървърна грешка: ${response.status}` }));
-            throw new Error(errorData.message || `Грешка от сървъра: ${response.status}`);
-        }
-        const result = await response.json();
-        if (!result.success || !result.showQuiz || !result.quizData) {
-            throw new Error(result.message || "Неуспешно зареждане или генериране на въпросник от сървъра.");
-        }
-        console.log("Quiz data received from worker (app.js):", result.quizData);
-        return result.quizData;
-    } catch (error) {
-        console.error("Error in generateAdaptiveQuizClientSide (app.js):", error);
-        throw error;
-    }
-}
-
-export async function _analyzeQuizAnswersAndAdaptClientSide(userId, quizId, submittedAnswers) { // Prefixed
-    console.log("analyzeQuizAnswersAndAdaptClientSide (app.js) called for quiz:", quizId, "answers:", submittedAnswers);
-    if (!userId || !quizId || !submittedAnswers) {
-        throw new Error("Липсват необходими данни за подаване на въпросника.");
-    }
-
-    if (userId.includes('test_user') || window.location.hostname.includes('replit')) {
-        console.log("Simulating quiz analysis for test user (app.js)");
-        return {
-            success: true,
-            message: "Тестовият въпросник (App.js) беше подаден успешно!",
-            aiUpdateSummary: { title: "Анализ (App.js) завършен", introduction: "Благодарим за отговорите!", changes: ["Актуализиран план според вашите нужди"], encouragement: "Продължавайте силната работа!" }
-        };
-    }
-
-    const payload = { userId: userId, quizId: quizId, answers: submittedAnswers };
-    try {
-        const response = await fetch(apiEndpoints.submitAdaptiveQuiz, {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify(payload)
-        });
-        if (!response.ok) {
-            const errorData = await response.json().catch(() => ({ message: `Сървърна грешка при подаване: ${response.status}` }));
-            throw new Error(errorData.message || `Грешка от сървъра при подаване на отговорите: ${response.status}`);
-        }
-        const result = await response.json();
-        if (!result.success) {
-            throw new Error(result.message || "Неуспешна обработка на отговорите от сървъра.");
-        }
-        console.log("Analysis result from worker (app.js):", result);
-        return result;
-    } catch (error) {
-        console.error("Error in analyzeQuizAnswersAndAdaptClientSide (app.js):", error);
-        throw error;
-    }
-}
-
-// Event handlers for quiz navigation, to be attached in eventListeners.js
-export function _handlePrevQuizQuestion() {
-    if (currentQuestionIndex > 0) {
-        setCurrentQuestionIndex(currentQuestionIndex - 1);
-        renderCurrentQuizQuestion(false); // from adaptiveQuiz.js
-    }
-}
-
-export function _handleNextQuizQuestion() {
-    const currentQ = currentQuizData.questions[currentQuestionIndex];
-    if (currentQ.required) {
-        const answer = userQuizAnswers[currentQ.id];
-        let isEmpty = false;
-        if (Array.isArray(answer)) { isEmpty = answer.length === 0; }
-        else { isEmpty = (answer === null || String(answer).trim() === ''); }
-
-        if (isEmpty) {
-            showQuizValidationMessage('Моля, отговорете на този въпрос, преди да продължите.'); // from adaptiveQuiz.js
-            const currentCard = selectors.quizQuestionContainer.querySelector('.aq-question-card-hybrid');
-            const inputAreaInCard = currentCard?.querySelector('.question-input-area');
-            const firstInputInArea = inputAreaInCard?.querySelector('input:not([type="hidden"]), textarea, .rating-square[tabindex="0"], select');
-            if (firstInputInArea) firstInputInArea.focus();
-            return;
-        }
-    }
-    hideQuizValidationMessage(); // from adaptiveQuiz.js
-    if (currentQuestionIndex < currentQuizData.questions.length - 1) {
-        setCurrentQuestionIndex(currentQuestionIndex + 1);
-        renderCurrentQuizQuestion(true); // from adaptiveQuiz.js
-    }
-}
-
-export async function _handleSubmitQuizAnswersClientSide() {
-    for (let i = 0; i < currentQuizData.questions.length; i++) {
-        const q = currentQuizData.questions[i];
-        if (q.required) {
-            const answer = userQuizAnswers[q.id];
-            let isEmpty = false;
-            if (Array.isArray(answer)) { isEmpty = answer.length === 0; }
-            else { isEmpty = (answer === null || String(answer).trim() === ''); }
-            if (isEmpty) {
-                showToast(`Моля, отговорете на всички задължителни въпроси. Въпрос "${q.text.substring(0,30)}..." е пропуснат.`, true, 4000);
-                setCurrentQuestionIndex(i);
-                renderCurrentQuizQuestion(true); // from adaptiveQuiz.js
-                return;
-            }
-        }
-    }
-
-    showLoading(true, "Обработка на вашите отговори...");
-    try {
-        const analysisResult = await _analyzeQuizAnswersAndAdaptClientSide(currentUserId, currentQuizData.quizId, userQuizAnswers);
-        showToast("Въпросникът е успешно подаден!", false, 2000);
-
-        setTimeout(() => {
-            // closeModal се импортира от uiHandlers.js и ще се използва директно
-            // ако selectors.adaptiveQuizModal е наличен.
-            // В adaptiveQuiz.js closeModal се извиква през genericCloseModal, което е alias на closeModal
-            if (selectors.adaptiveQuizModal) {
-                 closeModal('adaptiveQuizWrapper');
-            }
-        }, 1500);
-
-        const aiSummaryFromWorker = safeGet(analysisResult, 'aiUpdateSummary');
-        if (aiSummaryFromWorker && (aiSummaryFromWorker.title || aiSummaryFromWorker.introduction || aiSummaryFromWorker.changes || aiSummaryFromWorker.encouragement)) {
-            let summaryHtml = `<h3>${aiSummaryFromWorker.title || 'Резултат от Вашия Чек-ин'}</h3>`;
-            if (aiSummaryFromWorker.introduction) summaryHtml += `<p>${aiSummaryFromWorker.introduction.replace(/\n/g, '<br>')}</p>`;
-            if (aiSummaryFromWorker.changes && Array.isArray(aiSummaryFromWorker.changes) && aiSummaryFromWorker.changes.length > 0) {
-                summaryHtml += `<ul>${aiSummaryFromWorker.changes.map(ch => `<li>${String(ch).replace(/\n/g, '<br>')}</li>`).join('')}</ul>`;
-            }
-            if (aiSummaryFromWorker.encouragement) summaryHtml += `<p>${aiSummaryFromWorker.encouragement.replace(/\n/g, '<br>')}</p>`;
-
-            // openInfoModalWithDetails се импортира от uiHandlers
-            // и очаква key и type, които тук не са релевантни по същия начин.
-            // Затова директно задаваме съдържанието на infoModal.
-            if (selectors.infoModalTitle) selectors.infoModalTitle.textContent = aiSummaryFromWorker.title || 'Информация';
-            if (selectors.infoModalBody) selectors.infoModalBody.innerHTML = summaryHtml;
-            openModal('infoModal'); // openModal от uiHandlers
-
-            fetch(apiEndpoints.acknowledgeAiUpdate, {
-                 method: 'POST', headers: {'Content-Type': 'application/json'},
-                 body: JSON.stringify({userId: currentUserId})
-            }).catch(err => console.warn("Failed to acknowledge AI update to worker:", err));
-        } else if (analysisResult && analysisResult.message) {
-             showToast(analysisResult.message, false, 4000);
-        }
-
-        await loadDashboardData();
-        setCurrentQuizData(null);
-        setUserQuizAnswers({});
-        setCurrentQuestionIndex(0);
-    } catch (error) {
-        console.error("Error submitting quiz answers and adapting:", error);
-        showToast(`Грешка при подаване на въпросника: ${error.message}`, true);
-    } finally {
-        showLoading(false);
-    }
-}
-
-export async function _handleTriggerAdaptiveQuizClientSide() { // Exported for eventListeners.js
-    if (!currentUserId) { showToast("Моля, влезте първо.", true); return; }
-    _openAdaptiveQuizModal(); // from adaptiveQuiz.js
-}
-
-// ==========================================================================
-// ЧАТ ФУНКЦИИ (API комуникация и управление на историята)
-// ==========================================================================
-export async function handleChatSend() { // Exported for eventListeners.js
-    if (!selectors.chatInput || !selectors.chatSend) return;
-    const messageText = selectors.chatInput.value.trim();
-    if (!messageText || !currentUserId) return;
-
-    displayChatMessage(messageText, 'user'); // from chat.js
-    chatHistory.push({ text: messageText, sender: 'user', isError: false });
-
-    selectors.chatInput.value = ''; selectors.chatInput.disabled = true; selectors.chatSend.disabled = true;
-    displayChatTypingIndicator(true); // from chat.js
-    try {
-        const payload = { userId: currentUserId, message: messageText, history: chatHistory.slice(-10) }; // Send some history context
-        const response = await fetch(apiEndpoints.chat, {
-            method: 'POST', headers: { 'Content-Type': 'application/json' }, body: JSON.stringify(payload)
-        });
-        const result = await response.json();
-        if (!response.ok || !result.success) throw new Error(result.message || `HTTP ${response.status}`);
-
-        displayChatMessage(result.reply, 'bot'); // from chat.js
-        chatHistory.push({ text: result.reply, sender: 'bot', isError: false });
-
-    } catch (e) {
-        const errorMsg = `Грешка при комуникация с асистента: ${e.message}`;
-        displayChatMessage(errorMsg, 'bot', true); // from chat.js
-        chatHistory.push({ text: errorMsg, sender: 'bot', isError: true });
-    } finally {
-        displayChatTypingIndicator(false); // from chat.js
-        if(selectors.chatInput) { selectors.chatInput.disabled = false; selectors.chatInput.focus(); }
-        if(selectors.chatSend) selectors.chatSend.disabled = false;
-    }
-}
-export function handleChatInputKeypress(e){ if(e.key==='Enter'&&!e.shiftKey){e.preventDefault();handleChatSend();}} // Exported for eventListeners
 
 // ==========================================================================
 // СТАРТ НА ПРИЛОЖЕНИЕТО
